% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_treatments-data.R
\docType{data}
\name{infection_treatments}
\alias{infection_treatments}
\title{Infection treatments dataset}
\format{
A data frame with 150 rows and 2 variables:
\describe{
  \item{Infection}{Type of infection.}
  \item{Treatment}{Treatment administered.}
}
}
\source{
Created for the textbook *Reproducible Research Using R*.
}
\usage{
infection_treatments
}
\description{
A teaching dataset used in *Reproducible Research Using R*.
This dataset is used for examples involving categorical data analysis
and group comparisons.
}
\keyword{datasets}
