% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza_prices-data.R
\docType{data}
\name{pizza_prices}
\alias{pizza_prices}
\title{Pizza prices dataset}
\format{
A data frame with 156 rows and 3 variables:
\describe{
  \item{Week}{Week number of observation.}
  \item{Total Volume}{Total number of pizzas sold in that week.}
  \item{Total Price}{Total revenue generated from pizza sales in that week.}
}
}
\source{
Created for the textbook *Reproducible Research Using R*.
}
\usage{
pizza_prices
}
\description{
A teaching dataset used in *Reproducible Research Using R*.
This dataset is primarily used for examples involving data visualization
and linear regression.
}
\keyword{datasets}
