% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_chapters.R
\name{list_chapters}
\alias{list_chapters}
\title{List available chapter scripts included with the package}
\usage{
list_chapters()
}
\value{
A data frame with columns: chapter, type, file.
}
\description{
List available chapter scripts included with the package
}
\examples{
list_chapters()
}
